// ladder/web/src/gen/types/CreateAgent.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

import type { AgentBody } from './AgentBody'
import type { ErrorModel } from './ErrorModel'

export type CreateAgentStatus201 = AgentBody;

export type CreateAgentStatus409 = ErrorModel;

export type CreateAgentStatus422 = ErrorModel;

export type CreateAgentStatus500 = ErrorModel;

export const createAgentRequestKindEnum = {
    onnx: "onnx",
    builtin: "builtin"
} as const;

export type CreateAgentRequestKindEnumKey = (typeof createAgentRequestKindEnum)[keyof typeof createAgentRequestKindEnum];

export type CreateAgentRequestBody = {
    builtin_name?: string;
    /**
     * @description
     * Format: `binary`
     * @type string | undefined
    */
    file?: Blob;
    kind: CreateAgentRequestKindEnumKey;
    name: string;
} | undefined;

export type CreateAgentOptions = {
    body: CreateAgentRequestBody;
    path?: never;
    query?: never;
    headers?: never;
};

export type CreateAgentResponses = {
    "201": CreateAgentStatus201;
    "409": CreateAgentStatus409;
    "422": CreateAgentStatus422;
    "500": CreateAgentStatus500;
};

/**
 * @description Union of all possible responses
*/
export type CreateAgentResponse = (CreateAgentStatus201 | CreateAgentStatus409 | CreateAgentStatus422 | CreateAgentStatus500);