// ladder/web/src/gen/types/CreateGroupBody.ts
/**
* Generated by Kubb (https://kubb.dev/).
* Do not edit manually.
*/

export type CreateGroupBody = {
    /**
     * @description A URL to the JSON Schema for this object.
     *
     * Format: `uri`
     * @example /api/schemas/CreateGroupBody.json
     * @type string | undefined
    */
    readonly $schema?: string;
    name: string;
};